Initialised Playwright
Adds `ui-tests` with playwright testing framework, an express.js server to server some test html in the pipeline, and a new workflow which will run on pull requests.
Also alters the rust script to add a new envirnoment variable: `max_id` which will be used when generating the random ID (defaults to our favourite magic number). When used in conjunction with the new `test.db` we can generate an html page with a consistent content (we're testing the generation method; not the data!) which we can test using Playwright.

Tests to follow.

#42 - Adds the framework to allow consistent asset generation
#33 - Adds the foundations to enable UI Testing